add-remove-endpoint

+++ src/main/java/ru/yandex/practicum/filmorate/service/UserDbService.java
@@ -26,7 +26,10 @@ public class UserDbService {
     }
 
     public int deleteUser(int id) {
-        return userStorage.deleteUserById(id);
+        if (userStorage.isExistingUser(id))
+            return userStorage.deleteUserById(id);
+        else
+            throw new NotFoundException(String.format("Пользователь %d не может быть найден", id));
     }
 
     public List<User> getUsers() {
@@ -50,7 +53,10 @@ public class UserDbService {
     }
 
     public List<User> getFriends(int id) {
-        return userStorage.getFriendsById(id);
+        if (userStorage.isExistingUser(id))
+            return userStorage.getFriendsById(id);
+        else
+            throw new NotFoundException(String.format("Пользователь %d не может быть найден", id));
     }
 
     public List<User> getCommonFriends(int id, int otherId) {
